Attempt 2 - Replaced with floats

+++ src/a1/A1Novice.java
@@ -24,7 +24,7 @@ public class A1Novice {
 			int itemCounter = scan.nextInt();
 			
 			// Initializes the total to be 0 in the variable sum
-			double sum = 0;
+			float sum = 0;
 			
 			// This loop runs once for each *different* item in the customer's order
 			for (int k = 0; k < itemCounter; k++) {
@@ -36,13 +36,14 @@ public class A1Novice {
 				scan.next();
 				
 				// Initializes the total cost of each item to be its quantity multiplied by its price
-				double cost = itemQuantity * scan.nextDouble();
+				float cost = itemQuantity * scan.nextFloat();
 				
 				// Sums the total price of the order by adding the cost of each item as it is processed
 				sum = sum + cost;
 				
 				
 			}
+			
 			
 			// Prints to the console the first initial, last name, and order total in the format required
 			System.out.println(firstInitial + ". " + lastName + ":" + " " + sum);
